Rename api.http to api.app and update refs
Rename chat-agent/api/http.py to chat-agent/api/app.py and update all imports and references across the repo (Procfile, README, main.py, tests, and conftest). Adjust tests and patches to import from api.app and update related comments and fixtures. This is a straight module rename/refactor with no behavioral changes.
