+++ chat-agent/Procfile
@@ -1,1 +1,1 @@
-web: uvicorn api.http:app --host 0.0.0.0 --port ${PORT:-8000}
+web: uvicorn api.app:app --host 0.0.0.0 --port ${PORT:-8000}
